更新功能 1. 封禁管理 2.修复bug

+++ smart-common/src/main/java/cn/goroute/smart/common/entity/pojo/PostTag.java
@@ -29,6 +29,6 @@ public class PostTag extends BaseEntity implements Serializable {
 	/**
 	 * 
 	 */
-	private int tagUid;
+	private Long tagUid;
 
 }
